Restructure to restore boxmaker.py plugin entrypoint
