+++ README.md
@@ -155,3 +155,4 @@ version | Date | Notes
 0.99 | (4 June 2020) | Upgraded to support Inkscape v1.0, minor fixes and a tidy up of the parameters dialog layout
 1.0 |  (17 June 2020) | v1.0 final released: fixes and dogbone added - Mills now supported!
 1.1 |  (9 Aug 2021) | v1.1 with fixes for newer Inkscape versions - sorry for the delays
+1.2 | (18 Dec 2022) | v1.2 retructure as python package
