+++ pyproject.toml
@@ -6,7 +6,7 @@ requires-python = ">=3.7"
 license = "GPL-2.0"
 keywords = []
 authors = [
-  { name = "Paul Hutchison"},
+  {name = "Paul Hutchison"},
 ]
 classifiers = [
   "Development Status :: 4 - Beta",
@@ -19,7 +19,7 @@ classifiers = [
   "Programming Language :: Python :: Implementation :: CPython",
   "Programming Language :: Python :: Implementation :: PyPy",
 ]
-dependencies = ["inkex", "simplestyle"]
+dependencies = []
 dynamic = ["version"]
 
 [project.urls]
@@ -36,6 +36,7 @@ path = "tabbedboxmaker/__about__.py"
 
 [tool.hatch.envs.default]
 dependencies = [
+  "inkex",
   "pytest",
   "pytest-cov",
 ]
